==> 02.1.clinical_data/3.1.npi_data_analysis_and_table_for_lme.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.cm as cm
from matplotlib.gridspec import GridSpec
import matplotlib.dates as mdates

# import pyarrow as pa
# import pyarrow.parquet as pq

import seaborn as sns
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 13)

npi = pd.read_excel('npi_values_30_NOV_23.xlsx')
npi.columns
#'FrequencyXSeverity', 'thalias_total',

# number of data with correction
100-(len(npi[npi['type of date correction'].isna()])+len(npi[npi['type of date correction']=='keep original date']))*100/len(npi)

npi['type of date correction'].unique()

len(npi)

npi[['id', 'Event Name','datetime','amended datetime']]

#npi= npi[npi['amended datetime']!='-'].reset_index()

len(npi)

initial_dates = npi[['id','amended datetime']][npi['Event Name']=='Baseline']
initial_dates = initial_dates.rename(columns={'amended datetime':'new_baseline_date'}).reset_index()
npi = npi.merge(initial_dates, on=['id'],how='left')

npi['check_new_diff_with_baseline'] = npi.apply(lambda x:((pd.to_datetime(x['amended datetime']) - x['new_baseline_date']).days
                                                           if type(x['new_baseline_date'])!=float else '-'), axis=1)

#### to check the time difference that i insert as formula in excel, everything fine except few weir cells
#### npi[npi['new difference with timeline ']!=npi['check_new_diff_with_baseline']][['id','baseline_date', 'datetime','diff_with_baseline',  'type of date correction','new_baseline_date','amended datetime' ,
####      'new difference with timeline ','check_new_diff_with_baseline']]

npi['type of date correction'].unique()

maps= pd.read_excel('dyads_unprotected.xlsx')
maps['id'] = maps['Research ID'].apply(lambda x:x[:5])
npi = npi.merge(maps[['id','DYAD']], on='id',how='left')

npi[['id','DYAD','Event Name','baseline_date', 'datetime','diff_with_baseline',  'type of date correction','new_baseline_date','amended datetime',
     'new difference with timeline ']][npi['DYAD']==1021]

my_ids = npi.sort_values(by=['id'])['id'].unique()
len(my_ids)

#checking my score vs score
#npi[(npi['FrequencyXSeverity']==0)&(npi['thalias_total']!=0)]
#npi[(npi['FrequencyXSeverity']>0)]

fig, axs = plt.subplots(7, 10, figsize=(10*2.4, 7*3))
count_var=-1
fig.tight_layout(pad=2)

for j in range(0,7):
    for i in range(0,10):
        count_var = count_var+1
        if count_var<67:
            data=npi[npi['id']==my_ids[count_var]]
            clean_data = data[data['amended datetime']!='-']
#             axs[j,i].plot(data['datetime'], data['thalias_total'], 'o-', color='tomato',linewidth=5)
            axs[j,i].plot(clean_data['amended datetime'], clean_data['thalias_total'], 'o-', color='slateblue',linewidth=4, alpha=0.7)
            axs[j,i].set_ylim(-2, data['thalias_total'].max()+5)

            axs[j,i].text(0.5, 0.95, my_ids[count_var], color='black', horizontalalignment='center',
                  verticalalignment='center', transform=axs[j,i].transAxes, fontsize=12)
#             axs[j,i].text(0.5, 0.95, data['DYAD'].unique()[0], color='black', horizontalalignment='center',
#                   verticalalignment='center', transform=axs[j,i].transAxes, fontsize=12)
            axs[j,i].xaxis.set_major_formatter(mdates.DateFormatter("%Y-%b"))
            axs[j,i].tick_params(axis='x', rotation=45, labelsize=8)
            axs[j,i].spines[['top']].set_visible(False)

axs[0,5].text(0, 1.3, 'NPI Total Scores', color='black', horizontalalignment='center',
                  verticalalignment='center', transform=axs[0,5].transAxes, fontsize=24)
# axs[0,11].text(0, 1.33, 'Original Date', color='red', horizontalalignment='left',
#                   verticalalignment='center', transform=axs[0,11].transAxes, fontsize=20)
# axs[0,11].text(0, 1.2, 'Amended Date', color='blue', horizontalalignment='left',
#                   verticalalignment='center', transform=axs[0,11].transAxes, fontsize=20)
fig.savefig("npi_score_time.pdf", bbox_inches = "tight")

fig, axs = plt.subplots(7, 10, figsize=(10*2.4, 7*3))
count_var=-1
fig.tight_layout(pad=2)

for j in range(0,7):
    for i in range(0,10):
        count_var = count_var+1
        if count_var<67:
            data=npi[npi['id']==my_ids[count_var]]
            clean_data = data[data['amended datetime']!='-']
#             axs[j,i].plot(data['datetime'], data['thalias_total'], 'o-', color='tomato',linewidth=5)
            axs[j,i].plot(clean_data['amended datetime'], clean_data['sleep'], 'o-', color='slateblue',linewidth=4, alpha=0.7)
            axs[j,i].set_ylim(-2, data['sleep'].max()+5)

            axs[j,i].text(0.5, 0.95, my_ids[count_var], color='black', horizontalalignment='center',
                  verticalalignment='center', transform=axs[j,i].transAxes, fontsize=12)
#             axs[j,i].text(0.5, 0.95, data['DYAD'].unique()[0], color='black', horizontalalignment='center',
#                   verticalalignment='center', transform=axs[j,i].transAxes, fontsize=12)
            axs[j,i].xaxis.set_major_formatter(mdates.DateFormatter("%Y-%b"))
            axs[j,i].tick_params(axis='x', rotation=45, labelsize=8)
            axs[j,i].spines[['top']].set_visible(False)

axs[0,5].text(0, 1.3, 'NPI Sleep Scores', color='black', horizontalalignment='center',
                  verticalalignment='center', transform=axs[0,5].transAxes, fontsize=24)
# axs[0,11].text(0, 1.33, 'Original Date', color='red', horizontalalignment='left',
#                   verticalalignment='center', transform=axs[0,11].transAxes, fontsize=20)
# axs[0,11].text(0, 1.2, 'Amended Date', color='blue', horizontalalignment='left',
#                   verticalalignment='center', transform=axs[0,11].transAxes, fontsize=20)
fig.savefig("npi_sleep_score_time.pdf", bbox_inches = "tight")

br

npi = npi[npi['amended datetime']!='-']

npi[['id', 'Event Name']].groupby(by='id').count().mean()

#npi[['amended datetime','thalias_total','sleep']]

count = npi[['id','sleep']].groupby(by=['id']).count().reset_index()

plt.figure(figsize=(6,4))
plt.hist(count['sleep'], bins=range(0,14,1),rwidth=0.9, color='tomato')

#plt.yticks([2,4,9,14,19], rotation=0)
plt.xticks([0.5,1.5,2.5,3.5,4.5,5.5,6.5,7.5,8.5,9.5,10.5,11.5,12.5],
           [0,1,2,3,4,5,6,7,8,9,10,11,12], rotation=0)
plt.xlabel('Number Scores Per Participant')
plt.show()

#count.describe()

npi['thalias_total'].describe()

#npi['thalias_total'].describe()

plt.figure(figsize=(6,4))
plt.hist(npi['thalias_total'], bins=range(0,112,10),rwidth=0.9, color='tomato')

#plt.yticks([1,3,5,7,9], rotation=0)
plt.xticks([5,15,25,35,45,55,65,75,85,95,105],
           ['0-9','10-19','20-29','30-39','40-49','50-59','60-69','70-79','80-89','90-99','100-109'], rotation=15)
plt.xlabel('NPI Total Scores')
plt.show()

average = npi[['id','thalias_total']].groupby(by=['id']).mean().reset_index()

plt.figure(figsize=(6,4))
plt.hist(average['thalias_total'], bins=range(0,52,5),rwidth=0.9, color='tomato')

plt.yticks([2,4,9,14,19], rotation=0)
plt.xticks([2.5,7.5,12.5,17.5,22.5,27.5,32.5,37.5,42.5,47.5],
           ['0-5','5-10','10-15','15-20','20-25','25-30','30-35','35-40','40-45','45-50'], rotation=0)
plt.xlabel('NPI Total Score Average Per Participant')
plt.show()


#average.describe()

plt.figure(figsize=(6,4))
plt.hist(npi['sleep'], bins=range(0,14,1),rwidth=0.9, color='tomato')


#plt.yticks([1,3,5,7,9], rotation=0)
plt.xticks([0.5,1.5,2.5,3.5,4.5,5.5,6.5,7.5,8.5,9.5,10.5,11.5,12.5],
           [0,1,2,3,4,5,6,7,8,9,10,11,12], rotation=0)
plt.xlabel('NPI Sleep Scores')
plt.show()

#npi['sleep'].describe()

average = npi[['id','sleep']].groupby(by=['id']).mean().reset_index()

plt.figure(figsize=(6,4))
plt.hist(average['sleep'], bins=range(0,11,1),rwidth=0.9, color='tomato')

#plt.yticks([2,4,9,14,19], rotation=0)
plt.xticks([0.5,1.5,2.5,3.5,4.5,5.5,6.5,7.5,8.5,9.5],
           [0,1,2,3,4,5,6,7,8,9], rotation=0)
plt.xlabel('NPI Sleep Score Average Per Participant')
plt.show()


npi[['id', 'Event Name','datetime','amended datetime']]

#average.describe()

npi_lme = npi[['id', 'Event Name','datetime','type of date correction','amended datetime','thalias_total','thalias_total_frequency','thalias_total_sev',
     'Total Distress Score','sleep','freq_sleep', 'sev_sleep',
       'SLEEP:  Does the patient have difficulty sleeping (do not count as present if the patient simply gets up once or twice per night only to go to the bathroom and falls back asleep immediately)? Is he/she up at night? Does he/she wander at night, get dressed, or disturb your sleep? ',
       'Does the patient have difficulty falling asleep?)',
       'Does the patient get up during the night (do not count if the patient gets up once or twice per night only to go to the bathroom and falls back asleep immediately)?)',
       'Does the patient wander, pace, or get involved in inappropriate activities at night?)',
       'Does the patient awaken you during the night?)',
       'Does the patient awaken at night, dress, and plan to go out thinking that it is morning and time to start the day?)',
       'Does the patient awaken too early in the morning (earlier that was his/her habit)?)',
       'Does the patient sleep excessively during the day?)',
       'Does the patient have any other nighttime behaviours that bother you that we havent talked about?)','freq_delusions', 'sev_delusions', 'delusions',
       'freq_hallucinations', 'sev_hallucinations', 'hallucinations',
       'freq_agitation_aggression', 'sev_agitation_aggression',
       'agitation_aggression', 'freq_depression_dysphoria',
       'sev_depression_dysphoria', 'depression_dysphoria', 'freq_anxiety',
       'sev_anxiety', 'anxiety', 'freq_elation_euphoria',
       'sev_elation_euphoria', 'elation_euphoria', 'freq_apathy', 'sev_apathy',
       'apathy', 'freq_disinhibition', 'sev_disinhibition', 'disinhibition',
       'freq_irritability', 'sev_irritability', 'irritability',
       'freq_aberrant_motor_beh', 'sev_aberrant_motor_beh',
       'aberrant_motor_beh', 'freq_appetite', 'sev_appetite', 'appetite']]
#npi_lme.info()

npi_lme['amended month'] = pd.to_datetime(npi_lme['amended datetime']).dt.strftime('%B')#.month
npi_lme['amended bimonth'] ='JanFeb'
npi_lme.loc[(npi_lme['amended month']=='March')|(npi_lme['amended month']=='April'), 'amended bimonth'] = 'MarApr'
npi_lme.loc[(npi_lme['amended month']=='May')|(npi_lme['amended month']=='June'), 'amended bimonth'] = 'MayJun'
npi_lme.loc[(npi_lme['amended month']=='July')|(npi_lme['amended month']=='August'), 'amended bimonth'] = 'JulAug'
npi_lme.loc[(npi_lme['amended month']=='September')|(npi_lme['amended month']=='October'), 'amended bimonth'] = 'SepOct'
npi_lme.loc[(npi_lme['amended month']=='November')|(npi_lme['amended month']=='December'), 'amended bimonth'] = 'NovDec'


npi_lme['month'] = pd.to_datetime(npi_lme['datetime']).dt.strftime('%B')#.month
npi_lme['bimonth'] ='JanFeb'
npi_lme.loc[(npi_lme['month']=='March')|(npi_lme['month']=='April'), 'bimonth'] = 'MarApr'
npi_lme.loc[(npi_lme['month']=='May')|(npi_lme['month']=='June'), 'bimonth'] = 'MayJun'
npi_lme.loc[(npi_lme['month']=='July')|(npi_lme['month']=='August'), 'bimonth'] = 'JulAug'
npi_lme.loc[(npi_lme['month']=='September')|(npi_lme['month']=='October'), 'bimonth'] = 'SepOct'
npi_lme.loc[(npi_lme['month']=='November')|(npi_lme['month']=='December'), 'bimonth'] = 'NovDec'
npi_lme.head(3)

sorted_ids = pd.read_csv('sorted_ids_14_11_2023.csv')

## here by dropping nulls I limit to the participants in my data
npi_lme = npi_lme.merge(sorted_ids[['participant_id','min_date', 'max_date']], 
                                                                 left_on='id', right_on='participant_id', how='left')#.dropna()

npi_lme['days_after_min_date'] = (pd.to_datetime(npi_lme['amended datetime'],dayfirst=True)-pd.to_datetime(npi_lme['min_date'],dayfirst=True)).dt.days
npi_lme['closest_to_initial_day']=False
npi_lme['absolute_days_after_min_date'] = abs(npi_lme['days_after_min_date'])
npi_lme.loc[npi_lme.groupby('id', sort=False)['absolute_days_after_min_date'].idxmin(),'closest_to_initial_day']=True

npi_lme['days_after_max_date'] = (pd.to_datetime(npi_lme['amended datetime'],dayfirst=True)-pd.to_datetime(npi_lme['max_date'],dayfirst=True)).dt.days
npi_lme['closest_to_final_day']=False
npi_lme['absolute_days_after_max_date'] = abs(npi_lme['days_after_max_date'])
npi_lme.loc[npi_lme.groupby('id', sort=False)['absolute_days_after_max_date'].idxmin(),'closest_to_final_day']=True


npi_lme = npi_lme[['id', 'Event Name',  'amended datetime','closest_to_initial_day','days_after_min_date',
                   'closest_to_final_day', 'amended month', 'amended bimonth', 'participant_id', 'min_date', 'thalias_total',
                   'thalias_total_frequency', 'thalias_total_sev', 'Total Distress Score',
       'sleep', 'freq_sleep', 'sev_sleep',
       'SLEEP:  Does the patient have difficulty sleeping (do not count as present if the patient simply gets up once or twice per night only to go to the bathroom and falls back asleep immediately)? Is he/she up at night? Does he/she wander at night, get dressed, or disturb your sleep? ',
       'Does the patient have difficulty falling asleep?)',
       'Does the patient get up during the night (do not count if the patient gets up once or twice per night only to go to the bathroom and falls back asleep immediately)?)',
       'Does the patient wander, pace, or get involved in inappropriate activities at night?)',
       'Does the patient awaken you during the night?)',
       'Does the patient awaken at night, dress, and plan to go out thinking that it is morning and time to start the day?)',
       'Does the patient awaken too early in the morning (earlier that was his/her habit)?)',
       'Does the patient sleep excessively during the day?)',
       'Does the patient have any other nighttime behaviours that bother you that we havent talked about?)',
        'freq_delusions', 'sev_delusions', 'delusions',
       'freq_hallucinations', 'sev_hallucinations', 'hallucinations',
       'freq_agitation_aggression', 'sev_agitation_aggression',
       'agitation_aggression', 'freq_depression_dysphoria',
       'sev_depression_dysphoria', 'depression_dysphoria', 'freq_anxiety',
       'sev_anxiety', 'anxiety', 'freq_elation_euphoria',
       'sev_elation_euphoria', 'elation_euphoria', 'freq_apathy', 'sev_apathy',
       'apathy', 'freq_disinhibition', 'sev_disinhibition', 'disinhibition',
       'freq_irritability', 'sev_irritability', 'irritability',
       'freq_aberrant_motor_beh', 'sev_aberrant_motor_beh',
       'aberrant_motor_beh', 'freq_appetite', 'sev_appetite', 'appetite', 
                'type of date correction', 'datetime',  'month', 'bimonth']]
npi_lme 

npi_lme = npi_lme.replace(np.nan, '.', regex=True)
npi_lme = npi_lme.rename(columns={"sleep": "npi_sleep", "thalias_total": "npi_total",
                                 'thalias_total_frequency':'total_frequency', 'thalias_total_sev':'total_severity',
                                 'days_after_min_date':'time_in_study'})
 #       x['npi_sleep'] = npi.loc[found,'sleep'].values[0]
  #      x['npi_total'] = npi.loc[found,'thalias_total'].values[0]

npi_lme[npi_lme.isna().any(axis=1)]

npi_lme[['amended datetime','amended bimonth']].groupby('amended bimonth').count()

br

npi_lme[['datetime','bimonth']].groupby('bimonth').count()

#npi_lme[['amended datetime','amended month']].groupby('amended month').count()

#npi_lme[['datetime','month']].groupby('month').count()

#brnpi_lme.to_csv('npi_values_30_NOV_23_for_lme.csv')
